more boilerplate cleanup
