+++ package.json
@@ -24,6 +24,7 @@
     "body-parser": "^1.19.0",
     "borc": "^2.1.2",
     "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "env": "^0.0.2",
     "express": "^4.17.1",
