--- server/common/env.ts
@@ -1,2 +1,0 @@
-import dotenv from 'dotenv'
-dotenv.config()
