--- server/common/logger.ts
@@ -1,8 +1,0 @@
-import pino from 'pino'
-
-const l = pino({
-  name: process.env.APP_ID,
-  level: process.env.LOG_LEVEL,
-})
-
-export default l
