--- server/common/server.ts
@@ -1,55 +1,0 @@
-import express, { Application } from 'express'
-import path from 'path'
-import bodyParser from 'body-parser'
-import bearerToken from 'express-bearer-token'
-import http from 'http'
-import os from 'os'
-import cookieParser from 'cookie-parser'
-import l from './logger'
-
-import installValidator from './swagger'
-
-const app = express()
-
-export default class ExpressServer {
-  private routes: (app: Application) => void
-  constructor() {
-    const root = path.normalize(__dirname + '/../..')
-    app.use(bodyParser.json({ limit: process.env.REQUEST_LIMIT || '100kb' }))
-    app.use(
-      bodyParser.urlencoded({
-        extended: true,
-        limit: process.env.REQUEST_LIMIT || '100kb',
-      })
-    )
-    app.use(bodyParser.text({ limit: process.env.REQUEST_LIMIT || '100kb' }))
-    app.use(cookieParser(process.env.SESSION_SECRET))
-    app.use(bearerToken())
-    app.use(express.static(`${root}/public`))
-  }
-
-  router(routes: (app: Application) => void): ExpressServer {
-    this.routes = routes
-    return this
-  }
-
-  listen(port: number): Application {
-    const welcome = (p: number) => (): void =>
-      l.info(
-        `up and running in ${
-          process.env.NODE_ENV || 'development'
-        } @: ${os.hostname()} on port: ${p}}`
-      )
-
-    installValidator(app, this.routes)
-      .then(() => {
-        http.createServer(app).listen(port, welcome(port))
-      })
-      .catch((e) => {
-        l.error(e)
-        process.exit(1)
-      })
-
-    return app
-  }
-}
