+++ server/lib/auth.ts
