--- server/routes.ts
@@ -1,7 +1,0 @@
-import { Application } from 'express'
-import examplesRouter from './api/controllers/examples/router'
-import filecoinRouter from './api/controllers/filecoin/router'
-export default function routes(app: Application): void {
-  app.use('/api/v1/examples', examplesRouter)
-  app.use('/api/v1/filecoin', filecoinRouter)
-}
